@@ -8,8 +8,9 @@ using ServiceManager.ProjectDeployment.Data;
 using SharpRepository.Repository;
 using Tauron;
 using Tauron.Application.AkkaNode.Services;
-using Tauron.Application.AkkaNode.Services.Commands;
 using Tauron.Application.AkkaNode.Services.FileTransfer;
+using Tauron.Application.AkkaNode.Services.Reporting;
+using Tauron.Application.AkkaNode.Services.Reporting.Commands;
 using Tauron.Application.Files.VirtualFiles;
 using Tauron.Application.Master.Commands.Deployment.Build;
 using Tauron.Application.Master.Commands.Deployment.Build.Data;
