@@ -10,8 +10,9 @@ using ServiceManager.ProjectDeployment.Data;
 using Tauron;
 using Tauron.Akka;
 using Tauron.Application.AkkaNode.Services;
-using Tauron.Application.AkkaNode.Services.Commands;
 using Tauron.Application.AkkaNode.Services.FileTransfer;
+using Tauron.Application.AkkaNode.Services.Reporting;
+using Tauron.Application.AkkaNode.Services.Reporting.Commands;
 using Tauron.Application.Master.Commands.Deployment.Build;
 using Tauron.Application.Master.Commands.Deployment.Repository;
 using Tauron.Operations;
