@@ -1,6 +1,6 @@
 ﻿using JetBrains.Annotations;
 
-namespace Tauron.Application.AkkaNode.Services.Commands
+namespace Tauron.Application.AkkaNode.Services.Reporting.Commands
 {
     [PublicAPI]
     public sealed record FileTransactionId(string Id);
