@@ -5,7 +5,7 @@ using NLog;
 using Tauron.Akka;
 using Tauron.Host;
 
-namespace Tauron.Application.AkkaNode.Services.Commands
+namespace Tauron.Application.AkkaNode.Services.Reporting.Commands
 {
     public static class SendingHelper
     {
