@@ -1,4 +1,4 @@
-﻿namespace Tauron.Application.AkkaNode.Services
+﻿namespace Tauron.Application.AkkaNode.Services.Reporting
 {
     public interface IDelegatingMessage
     {
