@@ -2,7 +2,8 @@
 using Akka.Cluster.Tools.Singleton;
 using JetBrains.Annotations;
 using Tauron.Application.AkkaNode.Services;
-using Tauron.Application.AkkaNode.Services.Commands;
+using Tauron.Application.AkkaNode.Services.Reporting;
+using Tauron.Application.AkkaNode.Services.Reporting.Commands;
 
 namespace Tauron.Application.Master.Commands.Deployment.Repository
 {
