@@ -1,6 +1,6 @@
 ﻿using Akka.Cluster;
 
-namespace Tauron.Application.Master.Commands
+namespace Tauron.Application.Master.Commands.ServiceRegistry
 {
     public sealed record RegisterService(string Name, UniqueAddress Address);
 }
