@@ -6,7 +6,7 @@ using System.Reactive.Subjects;
 using System.Text;
 using System.Threading;
 using Newtonsoft.Json;
-using Serilog;
+using NLog;
 using Servicemnager.Networking;
 using Servicemnager.Networking.Data;
 using Servicemnager.Networking.IPC;
@@ -121,7 +121,7 @@ namespace WpfApp
                 IsReady = false;
                 ErrorMessage = e.Message;
 
-                Log.ForContext<IpcConnection>()
+                LogManager.GetCurrentClassLogger()
                     .Error(e, "Error on Starting Ipc");
 
                 Dispose();
